pyproject.toml
[project]
name = "uvb"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "biopython>=1.85",
    "ipykernel>=6.29.5",
]
